backend/token-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
cryptography==44.0.0
pycryptodome==3.21.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
